added notifications

+++ app/Http/Controllers/FifthYearController.php
@@ -5,10 +5,12 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\fifthyear;
-use Illuminate\Support\Arr;
+use App\User;
+use App\Notifications\FoundMatchNotification;
+use App\Notifications\DoubleSwitchNotification;
+use Ramsey\Collection\DoubleEndedQueue;
 
-
-class FifthYearController extends Controller
+class fifthyearController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -17,8 +19,14 @@ class FifthYearController extends Controller
      */
     public function index()
     {
-        return response()->json(fifthyear::all(),200);
+        $columns=['from','to'];
+        $fifthyear = DB::select('select * from fifthyear');
+        return response()->json([
+            'message' => 'your switch was stored successfully',
+            'data' => $fifthyear
+        ],200);
     }
+    
 
     /**
      * Show the form for creating a new resource.
@@ -45,14 +53,19 @@ class FifthYearController extends Controller
             'extra_courses' => 'required',
             'user_id' => 'required'
         ]);
+    
         $to = $request->input('to');
         $from = $request->input('from');
         $extra_courses = $request->input('extra_courses');
         $id = $request->input('user_id');
         $major = $request->input('major');
+   
         $match = DB::table('fifthyear')->where('major',$major)->where('extra_courses',$extra_courses) ->get();
-         $match= $match-> where('to',$from)-> orwhere('from',$to)->get() ;
+        
+        // $match= $match-> where('to',$from)-> orwhere('from',$to)->get() ;
+        
          $switchuser=null;
+        
          foreach($match as $user){
              if($user->from == $to && $user->to == $from){
                 $switchuser = [$user];
@@ -71,6 +84,7 @@ class FifthYearController extends Controller
              break;
              }
          }
+         
             if($switchuser == null){
                 $fifthyear = new fifthyear;
                 $fifthyear->user_id=$id;
@@ -78,6 +92,7 @@ class FifthYearController extends Controller
                 $fifthyear->from=$from;
                 $fifthyear->to=$to;
                 $fifthyear->extra_courses=$extra_courses;
+                
                 $fifthyear->save();
         return response()->json([
             'message' => 'your switch was stored successfully',
@@ -86,22 +101,38 @@ class FifthYearController extends Controller
     }
     else{
         $found=[];
+        $sender = User::where('id',$id)->first();
         foreach($switchuser as $deleted){
-            $this->destroy($deleted->id);
-            $user = DB::table('users')->where('id',$deleted->id);
-              $return =[
-            'email' => $user->personal_mail,
-            'phone' => $user->phone_num,
-            'name' => $user->name,
-            'uni-mail'=>$user->mail
-             ];
-             $found = Arr::prepend($found, $return);
+            $this->destroy($deleted->user_id);
+            $user = User::where('id',$deleted->user_id)->first();
+            // $return = $user->name;  
+                 
+            array_push($found, $user);
+           
         }
         
-        return response()->json([
-            'message' => 'found a match',
-            'match' => $found
-        ]);
+        if(count($found) == 1){
+          
+            $user->notify(new FoundMatchNotification($sender));
+            $sender->notify(new FoundMatchNotification($user)); 
+            return response()->json([
+                'message' => 'found a match',
+                'match' => $found
+            ]);
+        }
+        else{
+            $user1=$found[0];
+            $user2=$found[1];
+            $user1->notify(new DoubleSwitchNotification($sender,$user2));
+            $user2->notify(new DoubleSwitchNotification($sender,$user1));
+            $sender->notify(new DoubleSwitchNotification($user1,$user2)); 
+          return response()->json([
+                'message' => 'found a match',
+                'match' => $found
+            ]); 
+        }
+        
+       
 
     }
     }
@@ -137,8 +168,8 @@ class FifthYearController extends Controller
     public function update(Request $request, $id)
     {
         $fifthyear = fifthyear::findOrFail($id);
-        $fifthyear->update(request()->all);
-        return response()->json($fifthyear,201);
+        $fifthyear->update($request->all());
+        return response()->json($fifthyear, 200);
     }
 
     /**
@@ -149,8 +180,9 @@ class FifthYearController extends Controller
      */
     public function destroy($id)
     {
-        $fifthyear = fifthyear::findOrFail($id);
-        $fifthyear->delete();
-        return response()->json(null,204);
+ 
+        fifthyear::where('user_id',$id)->delete(); 
+         return response()->json(null,204);
+    
     }
 }
