+++ app/Http/Controllers/FourthYearController.php
@@ -5,10 +5,12 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\fourthyear;
-use Illuminate\Support\Arr;
+use App\User;
+use App\Notifications\FoundMatchNotification;
+use App\Notifications\DoubleSwitchNotification;
+use Ramsey\Collection\DoubleEndedQueue;
 
-
-class FourthYearController extends Controller
+class fourthyearController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -17,8 +19,14 @@ class FourthYearController extends Controller
      */
     public function index()
     {
-        return response()->json(fourthyear::all(),200);
+        $columns=['from','to'];
+        $fourthyear = DB::select('select * from fourthyear');
+        return response()->json([
+            'message' => 'your switch was stored successfully',
+            'data' => $fourthyear
+        ],200);
     }
+    
 
     /**
      * Show the form for creating a new resource.
@@ -45,14 +53,19 @@ class FourthYearController extends Controller
             'extra_courses' => 'required',
             'user_id' => 'required'
         ]);
+    
         $to = $request->input('to');
         $from = $request->input('from');
         $extra_courses = $request->input('extra_courses');
         $id = $request->input('user_id');
         $major = $request->input('major');
+   
         $match = DB::table('fourthyear')->where('major',$major)->where('extra_courses',$extra_courses) ->get();
-         $match= $match-> where('to',$from)-> orwhere('from',$to)->get() ;
+        
+        // $match= $match-> where('to',$from)-> orwhere('from',$to)->get() ;
+        
          $switchuser=null;
+        
          foreach($match as $user){
              if($user->from == $to && $user->to == $from){
                 $switchuser = [$user];
@@ -71,6 +84,7 @@ class FourthYearController extends Controller
              break;
              }
          }
+         
             if($switchuser == null){
                 $fourthyear = new fourthyear;
                 $fourthyear->user_id=$id;
@@ -78,6 +92,7 @@ class FourthYearController extends Controller
                 $fourthyear->from=$from;
                 $fourthyear->to=$to;
                 $fourthyear->extra_courses=$extra_courses;
+                
                 $fourthyear->save();
         return response()->json([
             'message' => 'your switch was stored successfully',
@@ -86,22 +101,38 @@ class FourthYearController extends Controller
     }
     else{
         $found=[];
+        $sender = User::where('id',$id)->first();
         foreach($switchuser as $deleted){
-            $this->destroy($deleted->id);
-            $user = DB::table('users')->where('id',$deleted->id);
-              $return =[
-            'email' => $user->personal_mail,
-            'phone' => $user->phone_num,
-            'name' => $user->name,
-            'uni-mail'=>$user->mail
-             ];
-             $found = Arr::prepend($found, $return);
+            $this->destroy($deleted->user_id);
+            $user = User::where('id',$deleted->user_id)->first();
+            // $return = $user->name;  
+                 
+            array_push($found, $user);
+           
         }
         
-        return response()->json([
-            'message' => 'found a match',
-            'match' => $found
-        ]);
+        if(count($found) == 1){
+          
+            $user->notify(new FoundMatchNotification($sender));
+            $sender->notify(new FoundMatchNotification($user)); 
+            return response()->json([
+                'message' => 'found a match',
+                'match' => $found
+            ]);
+        }
+        else{
+            $user1=$found[0];
+            $user2=$found[1];
+            $user1->notify(new DoubleSwitchNotification($sender,$user2));
+            $user2->notify(new DoubleSwitchNotification($sender,$user1));
+            $sender->notify(new DoubleSwitchNotification($user1,$user2)); 
+          return response()->json([
+                'message' => 'found a match',
+                'match' => $found
+            ]); 
+        }
+        
+       
 
     }
     }
@@ -136,11 +167,10 @@ class FourthYearController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $fourthyear = fourthyear::fondOrFail($id);
+        $fourthyear = fourthyear::findOrFail($id);
         $fourthyear->update($request->all());
-        return response()->json($fourthyear,201);
-        
-           }
+        return response()->json($fourthyear, 200);
+    }
 
     /**
      * Remove the specified resource from storage.
@@ -150,8 +180,9 @@ class FourthYearController extends Controller
      */
     public function destroy($id)
     {
-        $fourthyear = fourthyear::fondOrFail($id);
-        $fourthyear->delete();
-        return response()->json(null,204);
+ 
+        fourthyear::where('user_id',$id)->delete(); 
+         return response()->json(null,204);
+    
     }
 }
